moving graph3 to a more reasonable place
